Request parameters fetched independently of instance name
In #build_reaouce and #update the params were found by instance_name,
which required the instance name to have the same name as the controller
(including namespaces). This is no longer the case, it is now
automatically found from the controller's name, thus users are free to
modify the instance name, furthermore controller namespaces is much better
supported.