Please add support for API controllers.
Please add support to API controllers as well!

When I had to use API controllers with IR, I found that `inherit_resources` method was not available and found the source code was like this. Instead, I had to call the three lines in the method body. It would be great if the support was somehow built-in or at least documented. The API controllers worked exactly as I expected.