Update CI
Don't run lint as separate job since we have a shared, required, optimized workflow for Rubocop so warnings appear inline with PR changes. The spec lints should just be run as part of the standard tests. Ideally these would be updated to work based on the gemfile version in test matrix.