Make lockfile platforms easier to maintain
This should work equally well and has less churn when maintainers
upgrade their OS or switch platforms.

Recent versions of Bundler will "optimize" this, and as long as no
variants specific to the macOS version exist, it will lock only
arch + OS, but not the specific macOS version.

These should not get added in the future again as long as we use a
recent enough version of Bundler, and for this reason locked Bundler
has been bumped to 2.5.11

Ref: activeadmin/activeadmin#8344